+++ lib/protobuf/gen/rpc.pb.ex
@@ -526,6 +526,36 @@ defmodule ForgeAbi.ResponseGetTetherState do
   field :state, 2, type: ForgeAbi.TetherState
 end
 
+defmodule ForgeAbi.RequestGetSwapState do
+  @moduledoc false
+  use Protobuf, syntax: :proto3
+
+  @type t :: %__MODULE__{
+          address: String.t(),
+          keys: [String.t()],
+          height: non_neg_integer
+        }
+  defstruct [:address, :keys, :height]
+
+  field :address, 1, type: :string
+  field :keys, 2, repeated: true, type: :string
+  field :height, 3, type: :uint64
+end
+
+defmodule ForgeAbi.ResponseGetSwapState do
+  @moduledoc false
+  use Protobuf, syntax: :proto3
+
+  @type t :: %__MODULE__{
+          code: atom | integer,
+          state: ForgeAbi.SwapState.t() | nil
+        }
+  defstruct [:code, :state]
+
+  field :code, 1, type: ForgeAbi.StatusCode, enum: true
+  field :state, 2, type: ForgeAbi.SwapState
+end
+
 defmodule ForgeAbi.RequestStoreFile do
   @moduledoc false
   use Protobuf, syntax: :proto3
@@ -822,8 +852,12 @@ defmodule ForgeAbi.RequestGetConfig do
   @moduledoc false
   use Protobuf, syntax: :proto3
 
-  @type t :: %__MODULE__{}
-  defstruct []
+  @type t :: %__MODULE__{
+          parsed: boolean
+        }
+  defstruct [:parsed]
+
+  field :parsed, 1, type: :bool
 end
 
 defmodule ForgeAbi.ResponseGetConfig do
@@ -1153,6 +1187,40 @@ defmodule ForgeAbi.ResponseListTethers do
   field :code, 1, type: ForgeAbi.StatusCode, enum: true
   field :page, 2, type: ForgeAbi.PageInfo
   field :tethers, 3, repeated: true, type: ForgeAbi.TetherState
+end
+
+defmodule ForgeAbi.RequestListSwap do
+  @moduledoc false
+  use Protobuf, syntax: :proto3
+
+  @type t :: %__MODULE__{
+          paging: ForgeAbi.PageInput.t() | nil,
+          sender: String.t(),
+          receiver: String.t(),
+          available: boolean
+        }
+  defstruct [:paging, :sender, :receiver, :available]
+
+  field :paging, 1, type: ForgeAbi.PageInput
+  field :sender, 2, type: :string
+  field :receiver, 3, type: :string
+  field :available, 4, type: :bool
+end
+
+defmodule ForgeAbi.ResponseListSwap do
+  @moduledoc false
+  use Protobuf, syntax: :proto3
+
+  @type t :: %__MODULE__{
+          code: atom | integer,
+          page: ForgeAbi.PageInfo.t() | nil,
+          swap: [ForgeAbi.SwapState.t()]
+        }
+  defstruct [:code, :page, :swap]
+
+  field :code, 1, type: ForgeAbi.StatusCode, enum: true
+  field :page, 2, type: ForgeAbi.PageInfo
+  field :swap, 3, repeated: true, type: ForgeAbi.SwapState
 end
 
 defmodule ForgeAbi.RequestGetHealthStatus do